+++ showcase/screens/ListItems.tsx
@@ -5,7 +5,6 @@ import { ListItemCompact } from 'react-native-urbi-ui/components/ListItemCompact
 import { ListItemLarge } from 'react-native-urbi-ui/components/ListItemLarge';
 import { ListItemPlaceholder } from 'react-native-urbi-ui/components/ListItemPlaceholder';
 import { IconButtonRegular } from 'react-native-urbi-ui/molecules/buttons/iconButtons/IconButtonRegular';
-import { ChipAndLabel } from 'react-native-urbi-ui/molecules/content/ChipAndLabel';
 import { DoubleLabel } from 'react-native-urbi-ui/molecules/content/DoubleLabel';
 import { IconAndDoubleLabel } from 'react-native-urbi-ui/molecules/content/IconAndDoubleLabel';
 import { IconAndLabel } from 'react-native-urbi-ui/molecules/content/IconAndLabel';
@@ -15,6 +14,7 @@ import { EndDoubleLabelAndIcon } from 'react-native-urbi-ui/molecules/end/EndDou
 import { EndLabel } from 'react-native-urbi-ui/molecules/end/EndLabel';
 import { EndLabelAndIcon } from 'react-native-urbi-ui/molecules/end/EndLabelAndIcon';
 import { EndRealTime } from 'react-native-urbi-ui/molecules/end/EndRealTime';
+import { EndRealTimeDouble } from 'react-native-urbi-ui/molecules/end/EndRealTimeDouble';
 import { colors } from 'react-native-urbi-ui/utils/colors';
 import { showAlert, showLongAlert } from 'react-native-urbi-ui/utils/functions';
 import { placeholder, renderComponent } from '../utils/ComponentPreview';
@@ -70,10 +70,10 @@ class ListItems extends React.PureComponent<any> {
             content={
               <ChipOverLabel
                 chip={{ label: 'U3', color: colors.subway, icon: 'subway-small' }}
-                label="to WhateverStraße"
+                label="a very long label that should get cut at some point"
               />
             }
-            end={<EndRealTime label="2" subtitle="min" />}
+            end={<EndRealTimeDouble label="2" subtitle="min" />}
           />
         )}
         {renderComponent(
@@ -86,7 +86,32 @@ class ListItems extends React.PureComponent<any> {
                 bottomLabel="to WhateverStraße"
               />
             }
-            end={<EndRealTime label="5" subtitle="min" />}
+            end={<EndRealTimeDouble label="5" subtitle="min" />}
+          />
+        )}
+        {renderComponent(
+          'ListItemChip (w/ChipAndDoubleLabel and real time)',
+          <ListItemLarge
+            content={
+              <ChipAndDoubleLabel
+                chip={{ label: 'U3', color: colors.subway, icon: 'subway-small' }}
+                topLabel="No idea what goes here"
+                bottomLabel="a very long label that should get cut at some point"
+              />
+            }
+            end={<EndRealTimeDouble label="22" subtitle="min" />}
+          />
+        )}
+        {renderComponent(
+          'ListItemChip (w/ChipOverLabel and real time)',
+          <ListItemLarge
+            content={
+              <ChipOverLabel
+                chip={{ label: 'U3', color: colors.subway, icon: 'subway-small' }}
+                label="to WhateverStraße"
+              />
+            }
+            end={<EndRealTime label="2 min" />}
           />
         )}
         {renderComponent('ListItemPlaceholder', <ListItemPlaceholder lines={1} />)}