+++ src/molecules/end/EndRealTimeDouble.tsx
@@ -1,10 +1,15 @@
 import React from 'react';
 import { StyleSheet, Text, View, Image } from 'react-native';
 import { colors } from '../../utils/colors';
-import { EndLabelProps, endLabelStyles, labelStyle } from './EndLabel';
+import {
+  EndDoubleLabelProps,
+  endDoubleLabelStyles,
+  labelStyle,
+  subtitleStyle,
+} from './EndDoubleLabel';
 
 const styles = StyleSheet.create({
-  ...endLabelStyles,
+  ...endDoubleLabelStyles,
   FirstRow: {
     flexDirection: 'row',
     alignItems: 'flex-start',
@@ -19,18 +24,28 @@ const styles = StyleSheet.create({
 
 const realTimeIcon = require('../../assets/img/ic_realtime.png');
 
-const EndRealTimeUnmemoized = (props: EndLabelProps) => (
+const EndRealTimeDoubleUnmemoized = (props: EndDoubleLabelProps) => (
   <View style={[styles.Wrapper, props.style]}>
     <View style={styles.FirstRow}>
       <Image style={styles.RealTime} source={realTimeIcon} />
       <Text
-        style={props.textColor ? [labelStyle, { color: colors[props.textColor] }] : labelStyle}
+        style={props.labelColor ? [labelStyle, { color: colors[props.labelColor] }] : labelStyle}
         numberOfLines={1}
       >
         {props.label}
       </Text>
     </View>
+    <Text
+      style={
+        props.subtitleColor
+          ? [subtitleStyle, { color: colors[props.subtitleColor] }]
+          : subtitleStyle
+      }
+      numberOfLines={1}
+    >
+      {props.subtitle}
+    </Text>
   </View>
 );
 
-export const EndRealTime = React.memo(EndRealTimeUnmemoized);
+export const EndRealTimeDouble = React.memo(EndRealTimeDoubleUnmemoized);